Enable balo tests

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/writer/BIRBinaryWriter.java
@@ -327,7 +327,7 @@ public class BIRBinaryWriter {
                 break;
             case TypeTags.BYTE:
                 int byteValue = ((Number) constValue.value).intValue();
-                buf.writeByte(addByteCPEntry(byteValue));
+                buf.writeInt(addByteCPEntry(byteValue));
                 break;
             case TypeTags.FLOAT:
                 // TODO:Remove the instanceof check by converting the float literal instance in Semantic analysis phase
